Bootloader Updater: Support product and vendor
If built as a product target, the updates will be read from product. The
pre-existing target will continue to work on system as before.

Vendor is also added as a target. This is only intended to support a
non-system partition for products without a product partition. If
product is available it is recommended to use that instead.

Change-Id: I4c86756bf4bd56cb9d365729b95c9b9860fc33e5